Unschedule alarms at disabling and deletion

+++ app/src/main/java/com/sbcf/pillbox/features/medicationreminders/screens/MedicationReminderEditScreen.kt
@@ -45,7 +45,7 @@ import com.sbcf.pillbox.utils.Modifiers.scaffoldedContent
 fun EditMedicationReminderScreen(
     vm: MedicationReminderEditViewModel = hiltViewModel(),
     onBackClick: () -> Unit,
-    reminderId: Long? = null
+    reminderId: Int? = null
 ) {
     LaunchedEffect(key1 = vm, key2 = reminderId) {
         vm.fetchReminder(reminderId)
